Fix race condition between belts and plot_graphs
The belt_shaper_calibration and axes_shaper_calibration macros are implemented such that the underlying klipper macros write raw data to /tmp.  However, klipper generally creates a file in /tmp before writing data to it.  This can result in a race condition, where plot_graphs.sh can find the file and move it out of /tmp before klipper has finished writing data to the file.

This commit resolves this race condition by implementing a check to see if the file is open using the linux "lsof" utility.