Update overrides.cfg
Removed PID's and z_offset to move them to printer.cfg as these are not overrides, and are base configurations.  While most of the probes the XY will be very close to what they actually are, X and Y doesn't need the exact precision as the Z offset and thus a close baseline config included in the /config/hardware/probes/ configs will work just fine.  If a user has a probe that is drastically different than what is the baseline, these would be overrides.

Additionally, the PID's are base configurations and not overrides.