Update printer.cfg
Moved z_offset and PID's from overrides.cfg as these configs are not overrides from other configurations and are baseline configs to be selected.  Example of this would be similar to the inductive.cfg, where the X&y offset are standard.  If the user has a non standard inductive probe, the x_offset & y_offset would be overrides and need to be configured in overrides.cfg.  

Uncommenting the PID's and z_offset in the printer.cfg would be no different than uncommenting the configuration file needed to operate, but will give the added benefit of the save_config command to work to update these user required settings.