Add Cartographer3D V3 with Touch configuration
We need to be able to override some settings for
`[bed_mesh]` so the include in `user_templates/printer.cfg`
is placed after the bed mesh definitions and not
in the probe section.

Cartographer3D V3 doesn't touch the bed with the nozzle
except for the `CARTOGRAPHER_TOUCH` command. Therefore
we need to pay attention to the nozzle temperature
for that. But not for all the other actions like
homing or quad gantry leveling or some such.

The documentation for Cartographer3D advises against
an adaptive bed mesh and the full bed mesh will
be fast enough in most cases, so we deactivated
the adaptive bed mesh if a Cartographer3D probe
is used.